Replace active and remove unused sarif functions
