+++ pkg/report/template.go
@@ -14,8 +14,6 @@ import (
 
 	"github.com/Masterminds/sprig"
 	"golang.org/x/xerrors"
-
-	"github.com/aquasecurity/trivy-db/pkg/vulnsrc/vulnerability"
 )
 
 // regex to extract file path in case string includes (distro:version)
@@ -46,18 +44,6 @@ func NewTemplateWriter(output io.Writer, outputTemplate string) (*TemplateWriter
 		}
 		return escaped.String()
 	}
-	templateFuncMap["makeRuleMap"] = func() map[string]int {
-		return make(map[string]int)
-	}
-	templateFuncMap["indexRule"] = func(rules map[string]int, vulnerabilityID string) bool {
-		if _, ok := rules[vulnerabilityID]; !ok {
-			rules[vulnerabilityID] = len(rules)
-			return true
-		}
-		return false
-	}
-	templateFuncMap["toSarifErrorLevel"] = toSarifErrorLevel
-	templateFuncMap["toSarifRuleName"] = toSarifRuleName
 	templateFuncMap["endWithPeriod"] = func(input string) string {
 		if !strings.HasSuffix(input, ".") {
 			input += "."
@@ -69,14 +55,6 @@ func NewTemplateWriter(output io.Writer, outputTemplate string) (*TemplateWriter
 	}
 	templateFuncMap["escapeString"] = func(input string) string {
 		return html.EscapeString(input)
-	}
-	templateFuncMap["toPathUri"] = func(input string) string {
-		var matches = re.FindStringSubmatch(input)
-		if matches != nil {
-			input = matches[re.SubexpIndex("path")]
-		}
-		input = strings.ReplaceAll(input, "\\", "/")
-		return input
 	}
 	templateFuncMap["getEnv"] = func(key string) string {
 		return os.Getenv(key)
@@ -98,31 +76,4 @@ func (tw TemplateWriter) Write(report Report) error {
 		return xerrors.Errorf("failed to write with template: %w", err)
 	}
 	return nil
-}
-
-func toSarifRuleName(vulnerabilityType string) string {
-	switch vulnerabilityType {
-	case vulnerability.Ubuntu, vulnerability.Alpine, vulnerability.RedHat, vulnerability.RedHatOVAL,
-		vulnerability.Debian, vulnerability.DebianOVAL, vulnerability.Fedora, vulnerability.Amazon,
-		vulnerability.OracleOVAL, vulnerability.SuseCVRF, vulnerability.OpenSuseCVRF, vulnerability.Photon,
-		vulnerability.CentOS:
-		return "OsPackageVulnerability"
-	case "npm", "yarn", "nuget", "pipenv", "poetry", "bundler", "cargo", "composer":
-		return "ProgrammingLanguageVulnerability"
-	default:
-		return "OtherVulnerability"
-	}
-}
-
-func toSarifErrorLevel(severity string) string {
-	switch severity {
-	case "CRITICAL", "HIGH":
-		return "error"
-	case "MEDIUM":
-		return "warning"
-	case "LOW", "UNKNOWN":
-		return "note"
-	default:
-		return "none"
-	}
 }
